# openapi/backend/openapi.yaml
...
openapi: '3.0.4'
info:
  title: webstatus.dev API
  version: 0.1.0
  description: >
    A tool to monitor and track the status of all Web Platform features across
    dimensions that are related to availability and implementation quality
    across browsers, and adoption by web developers.
servers:
  - url: https://api.webstatus.dev
    description: Production
paths:
  /v1/features:
    get:
      summary: List features
      operationId: listFeatures
      parameters:
        - $ref: '#/components/parameters/paginationTokenParam'
        - $ref: '#/components/parameters/paginationSizeParam'
        - in: query
          name: wpt_metric_view
          schema:
            $ref: '#/components/schemas/WPTMetricView'
        - in: query
          name: q
          description: >
            A query string to represent the filters to apply the datastore while searching.
            The query must follow the ANTLR grammar. Please read the query readme at antlr/FeatureSearch.md.
            The query must be url safe.
          required: false
          schema:
            type: string
            minLength: 1
        - in: query
          name: sort
          description: >
            Field to sort by, with 'asc' for ascending and 'desc' for descending order.
            Defaults to sorting by 'name' in ascending order (e.g., 'name_asc').
          required: false
          schema:
            type: string
            enum:
              - name_asc
              - name_desc
              - baseline_status_asc
              - baseline_status_desc
              - stable_chrome_asc
              - stable_chrome_desc
              - stable_safari_asc
              - stable_safari_desc
              - stable_edge_asc
              - stable_edge_desc
              - stable_firefox_asc
              - stable_firefox_desc
              - stable_chrome_android_asc
              - stable_chrome_android_desc
              - stable_firefox_android_asc
              - stable_firefox_android_desc
              - stable_safari_ios_asc
              - stable_safari_ios_desc
              - experimental_chrome_asc
              - experimental_chrome_desc
              - experimental_safari_asc
              - experimental_safari_desc
              - experimental_edge_asc
              - experimental_edge_desc
              - experimental_firefox_asc
              - experimental_firefox_desc
              - experimental_chrome_android_asc
              - experimental_chrome_android_desc
              - experimental_firefox_android_asc
              - experimental_firefox_android_desc
              - experimental_safari_ios_asc
              - experimental_safari_ios_desc
              - chrome_usage_asc
              - chrome_usage_desc
              - developer_signal_upvotes_asc
              - developer_signal_upvotes_desc
              # availability_{browser}_{direction}
              - availability_chrome_asc
              - availability_chrome_desc
              - availability_edge_asc
              - availability_edge_desc
              - availability_firefox_asc
              - availability_firefox_desc
              - availability_safari_asc
              - availability_safari_desc
              - availability_chrome_android_asc
              - availability_chrome_android_desc
              - availability_firefox_android_asc
              - availability_firefox_android_desc
              - availability_safari_ios_asc
              - availability_safari_ios_desc
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeaturePage'
        '400':
          description: Bad Input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicErrorModel'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicErrorModel'
        '429':
          description: Rate Limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicErrorModel'
        '500':
          description: Internal Service Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicErrorModel'
  /v1/features/{feature_id}:
    parameters:
      - name: feature_id
        in: path
        description: Feature ID
        required: true
        schema:
          type: string
      - in: query
        name: wpt_metric_view
        schema:
          $ref: '#/components/schemas/WPTMetricView'
    get:
      summary: Get Feature
      operationId: getFeature
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Feature'
        '301':
          description: Moved Permanently
          headers:
            Location:
              description: The URL of the new, permanent location for the feature.
              schema:
                type: string
                format: uri
        '400':
          description: Bad Input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicErrorModel'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicErrorModel'
        '410':
          description: Gone (Feature is no longer available at this URI due to evolution)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureGoneError'
        '429':
          description: Rate Limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicErrorModel'
        '500':
          description: Internal Service Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicErrorModel'
...
  /v1/features/{feature_id}/stats/wpt/browsers/{browser}/channels/{channel}/{metric_view}:
    parameters:
      - name: feature_id
        in: path
        description: Feature ID
        required: true
        schema:
          type: string
      - $ref: '#/components/parameters/browserPathParam'
      - $ref: '#/components/parameters/channelPathParam'
      - $ref: '#/components/parameters/metricViewPathParam'
    get:
      summary: Retrieve the wpt stats for a particular feature.
      operationId: listFeatureWPTMetrics
      parameters:
        - $ref: '#/components/parameters/startAtParam'
        - $ref: '#/components/parameters/endAtParam'
        - $ref: '#/components/parameters/paginationTokenParam'
        - $ref: '#/components/parameters/paginationSizeParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WPTRunMetricsPage'
        '400':
          description: Bad Input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicErrorModel'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicErrorModel'
        '429':
          description: Rate Limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicErrorModel'
        '500':
          description: Internal Service Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicErrorModel'
...
  /v1/stats/features/browsers/{browser}/feature_counts:
    parameters:
      - $ref: '#/components/parameters/browserPathParam'
    get:
      summary: >
        Returns the count of features supported for a specified browser over time.
        The timestamps for the individual metrics represent the releases of the
        specified browser.
      operationId: listAggregatedFeatureSupport
      parameters:
        - $ref: '#/components/parameters/startAtParam'
        - $ref: '#/components/parameters/endAtParam'
        - $ref: '#/components/parameters/paginationTokenParam'
        - $ref: '#/components/parameters/paginationSizeParam'
        - $ref: '#/components/parameters/includeBaselineMobileBrowsersParam'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrowserReleaseFeatureMetricsPage'
        '400':
          description: Bad Input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicErrorModel'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicErrorModel'
        '429':
          description: Rate Limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicErrorModel'
        '500':
          description: Internal Service Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicErrorModel'
...
  /v1/stats/features/browsers/{browser}/missing_one_implementation_counts/{date}/features:
    get:
      summary: >
        Returns the list of feature IDs that were counted as missing one
        implementation on a specific date for the given browser.
      operationId: listMissingOneImplementationFeatures
      parameters:
        - $ref: '#/components/parameters/browserPathParam'
        - in: path
          name: date
          schema:
            type: string
            format: date
          required: true
          description: The date for which to retrieve the feature IDs.
        - $ref: '#/components/parameters/paginationTokenParam'
        - $ref: '#/components/parameters/paginationSizeParam'
        - $ref: '#/components/parameters/includeBaselineMobileBrowsersParam'
        - in: query # Add otherBrowsers parameter
          name: browser
          description: >
            A list of browsers to check if {browser} is lagging behind on
            features. The list is provided by specifying repeating query parameters.
            Example:?browser=fooBrowser&browser=bazBrowser
          required: true
          schema:
            type: array
            items:
              $ref: '#/components/schemas/SupportedBrowsers'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MissingOneImplFeaturesPage'
        '400':
          description: Bad Input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicErrorModel'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicErrorModel'
        '429':
          description: Rate Limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicErrorModel'
        '500':
          description: Internal Service Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicErrorModel'
...
components:
  parameters:
    browserPathParam:
      in: path
      name: browser
      description: Browser name
      required: true
      schema:
        $ref: '#/components/schemas/SupportedBrowsers'
    channelPathParam:
      in: path
      name: channel
      description: Browser name
      required: true
      schema:
        type: string
        # List of supported channels that webstatus.dev currently ingests
        enum:
          - stable
          - experimental
    metricViewPathParam:
      in: path
      name: metric_view
      description: Specified metric view of the WPT data.
      required: true
      schema:
        $ref: '#/components/schemas/WPTMetricView'
    # TODO: Create a start_at version of this parameter
    startAtParam:
      in: query
      name: startAt
      schema:
        type: string
        format: date
      description: Start date (RFC 3339, section 5.6, for example, 2017-07-21). The date is inclusive.
      required: true
    # TODO: Create a end_at version of this parameter
    endAtParam:
      in: query
      name: endAt
      schema:
        type: string
        format: date
      description: End date (RFC 3339, section 5.6, for example, 2017-07-21). The date is exclusive.
      required: true
    paginationTokenParam:
      in: query
      name: page_token
      schema:
        type: string
      description: Pagination token
      required: false
    paginationSizeParam:
      in: query
      name: page_size
      schema:
        type: integer
        maximum: 100
        minimum: 1
        default: 1
      required: false
      description: Number of results to return
    includeBaselineMobileBrowsersParam:
      in: query
      name: include_baseline_mobile_browsers
      schema:
        type: boolean
      description: >
        Include baseline mobile browsers in the response. This is only applicable
        for desktop browsers with a corresponding mobile browser.
      required: false
  schemas:
    SupportedBrowsers:
      type: string
      description: List of supported browsers that webstatus.dev currently ingests
      enum:
        - chrome
        - firefox
        - safari
        - edge
        - chrome_android
        - firefox_android
        - safari_ios
    WPTMetricView:
      type: string
      description: The desired view of the WPT Data
      enum:
        - test_counts
        - subtest_counts
    BrowserReleaseFeatureMetric:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: >
            The timestamp that correlates to an event that may influence the
            count of features for a browser. This may be the release of the
            browser itself (when a browser may support a new set of features),
            or the release of another browser (when another browser supports
            a feature but our specified browser is now lagging behind with a new
            feature). Refer to the individual endpoint for more context on the
            use of this component.
        count:
          type: integer
          description: Total count of features.
          format: int64
      required:
        # For now, only require timestamp in case the definition
        # of the metric needs to change. Similar to WPTRunMetric.
        - timestamp
    BrowserReleaseFeatureMetricsPage:
      type: object
      properties:
        metadata:
          $ref: '#/components/schemas/PageMetadata'
        data:
          type: array
          items:
            $ref: '#/components/schemas/BrowserReleaseFeatureMetric'
      required:
        - data
...
    WPTRunMetric:
      type: object
      properties:
        run_timestamp:
          type: string
          format: date-time
          description: The start timestamp of the run.
        total_tests_count:
          type: integer
          description: Total number of tests
          format: int64
        test_pass_count:
          type: integer
          description: Number of passing tests
          format: int64
      required:
        - run_timestamp
...
    WPTRunMetricsPage:
      type: object
      properties:
        metadata:
          $ref: '#/components/schemas/PageMetadata'
        data:
          type: array
          items:
            $ref: '#/components/schemas/WPTRunMetric'
      required:
        - data
...
    WPTFeatureData:
      type: object
      properties:
        score:
          type: number
          format: double
        metadata:
          type: object
          description: >
            Contains optional metadata about the metric. This key-value pair is to be considered unstable and can change
            at any moment. If a field here becomes mature, we can add it to the main WPTFeatureData definition.
          additionalProperties: true
    BrowserImplementation:
      type: object
      properties:
        status:
          type: string
          enum:
            - unavailable
            - available
          # Go enum names
          x-enum-varnames:
            - Unavailable
            - Available
        date:
          type: string
          format: date
          description: The date on which the feature was implemented in a browser. (RFC 3339, section 5.6, for example, 2017-07-21)
        version:
          type: string
          description: The browser version in which the feature became available.
    PageMetadata:
      type: object
      properties:
        next_page_token:
          type: string
    PageMetadataWithTotal:
      type: object
      properties:
        next_page_token:
          type: string
        total:
          type: integer
          format: int64 # Currently, 'count(*)' from spanner comes as int64 https://cloud.google.com/spanner/docs/reference/standard-sql/data-types
      required:
        - total
    FeatureDeveloperSignals:
      type: object
      properties:
        upvotes:
          type: integer
          description: Count of positive feedback for a feature
          format: int64 # To match the value type from Spanner as it only supports int64.
        link:
          type: string
          description: Link to the issue tracking developer signal for a feature.
    FeatureDiscouragedInfo:
      type: object
      properties:
        according_to:
          type: array
          items:
            $ref: '#/components/schemas/FeatureDiscouragedAccordingTo'
        alternatives:
          type: array
          items:
            $ref: '#/components/schemas/FeatureDiscouragedAlternative'
    FeatureDiscouragedAccordingTo:
      type: object
      properties:
        link:
          type: string
      required:
        - link
    FeatureDiscouragedAlternative:
      type: object
      properties:
        id:
          type: string
      required:
        - id
    FeaturePage:
      type: object
      properties:
        metadata:
          $ref: '#/components/schemas/PageMetadataWithTotal'
        data:
          type: array
          items:
            $ref: '#/components/schemas/Feature'
      required:
        - data
        - metadata
    FeatureWPTSnapshots:
      type: object
      properties:
        stable:
          type: object
          description: >
            Contains snapshot of the stable WPT data. The keys for the
            object comes from the different cases in
            https://github.com/web-platform-tests/wpt.fyi/blob/fb5bae7c6d04563864ef1c28a263a0a8d6637c4e/shared/product_spec.go#L71-L104
          additionalProperties:
            $ref: '#/components/schemas/WPTFeatureData'
        experimental:
          type: object
          description: >
            Contains snapshot of the experimental WPT data. The keys for the
            object comes from the different cases in
            https://github.com/web-platform-tests/wpt.fyi/blob/fb5bae7c6d04563864ef1c28a263a0a8d6637c4e/shared/product_spec.go#L71-L104
          additionalProperties:
            $ref: '#/components/schemas/WPTFeatureData'
    BaselineInfo:
      type: object
      description: Contains baseline information for a feature.
      properties:
        status:
          type: string
          enum:
            - limited
            - newly
            - widely
        low_date:
          type: string
          format: date
        high_date:
          type: string
          format: date
    FeatureSpecInfo:
      type: object
      properties:
        links:
          type: array
          items:
            $ref: '#/components/schemas/SpecLink'
    SpecLink:
      type: object
      properties:
        link:
          type: string
...
    VendorPosition:
      type: object
      description: >
        A loosely defined object representing a single vendor's standards position.
        The schema is intentionally open-ended (`additionalProperties: true`) to
        accommodate the evolving structure of the upstream web-features-mappings data source.
      additionalProperties: true
    Feature:
      type: object
      properties:
        feature_id:
          type: string
        system_managed_saved_search_id:
          type: string
          nullable: true
          readOnly: true
          description: The ID of the system-managed saved search for this feature.
        name:
          type: string
          description: >
            Short name that is intended to be human friendly.
            Comes from FeatureData's 'name' field in
            https://github.com/web-platform-dx/web-features/blob/main/schemas/data.schema.json
        spec:
          $ref: '#/components/schemas/FeatureSpecInfo'
        browser_implementations:
          type: object
          description: >
            Describes the implementation status of the feature. The keys for the object come
            from https://github.com/web-platform-dx/web-features/blob/f630e000a15eaff23f41d6b8c4a44cca64a4a7e4/schemas/data.schema.json#L261-L289
          additionalProperties:
            $ref: '#/components/schemas/BrowserImplementation'
        baseline:
          $ref: '#/components/schemas/BaselineInfo'
        usage:
          $ref: '#/components/schemas/BrowserUsage'
        wpt:
          $ref: '#/components/schemas/FeatureWPTSnapshots'
        developer_signals:
          description: >
            The absence of this object means the feature does not have an active issue
            in https://github.com/web-platform-dx/developer-signals.
          $ref: '#/components/schemas/FeatureDeveloperSignals'
        discouraged:
          $ref: '#/components/schemas/FeatureDiscouragedInfo'
        vendor_positions:
          type: array
          description: >
            Contains standards positions from various vendors.
          items:
            $ref: '#/components/schemas/VendorPosition'
      required:
        - feature_id
        - name
        - baseline_status
    FeatureSplitInfo:
      type: object
      properties:
        id:
          type: string
          description: The ID of a feature that resulted from a split.
      required:
        - id
...
    FeatureGoneError:
      description: |
        Represents various reasons a feature might be gone due to evolution.
        Clients should check the 'type' field to determine the specific error structure.
      $ref: '#/components/schemas/FeatureGoneSplit'
      # Cannot use oneOf until this bug is fixed
      # https://github.com/oapi-codegen/oapi-codegen/issues/970
      # oneOf:
      #   - $ref: '#/components/schemas/FeatureGoneSplit'
      # discriminator:
      #   propertyName: type
      #   mapping:
      #     split: '#/components/schemas/FeatureGoneSplit'
    FeatureGoneSplit:
      allOf:
        - $ref: '#/components/schemas/BasicErrorModel'
        - type: object
          properties:
            type:
              type: string
              enum:
                - split
              description: Indicates the feature was removed due to a split.
            new_features:
              type: array
              description: >
                A list of feature IDs that have completely replaced the original,
                now-gone feature, as a result of a full split.
              items:
                $ref: '#/components/schemas/FeatureSplitInfo'
          required:
            - type
            - new_features
    BrowserUsage:
      type: object
      properties:
        chrome:
          $ref: '#/components/schemas/ChromeUsageInfo'
    ChromeUsageInfo:
      type: object
      properties:
        daily:
          type: number
          format: double
          description: >
            Latest snapshot of the usage metric for the given feature.
          minimum: 0.0
          maximum: 100.0
    BasicErrorModel:
      type: object
      required:
        - message
        - code
      properties:
        message:
          type: string
        code:
          type: integer
          minimum: 100
          maximum: 600
...
    MissingOneImplFeature: # Schema for a single feature object
      type: object
      properties:
        feature_id:
          type: string
          description: The ID of a feature.
    MissingOneImplFeaturesPage: # Schema for the paginated response
      type: object
      properties:
        metadata:
          $ref: '#/components/schemas/PageMetadata'
        data:
          type: array
          items:
            $ref: '#/components/schemas/MissingOneImplFeature'
      required:
        - data
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    noAuth:
      description: No authentication required. User can opt out of providing a bearer token.
      type: apiKey
      in: header
      name: X-No-Auth